Address code review comments

+++ mlir-sycl/lib/Dialect/SYCL/Analysis/MemoryAccessAnalysis.cpp
@@ -82,7 +82,7 @@ static bool isSmallerThanNegativeOne(Value val, DataFlowSolver &solver) {
 // MemoryAccessMatrix
 //===----------------------------------------------------------------------===//
 
-MemoryAccessMatrix::MemoryAccessMatrix(unsigned nRows, unsigned nColumns)
+MemoryAccessMatrix::MemoryAccessMatrix(size_t nRows, size_t nColumns)
     : nRows(nRows), nColumns(nColumns), data(nRows * nColumns) {}
 
 MemoryAccessMatrix::MemoryAccessMatrix(
@@ -103,109 +103,109 @@ MemoryAccessMatrix::MemoryAccessMatrix(
     data.append(initRow);
 }
 
-void MemoryAccessMatrix::swapRows(unsigned row, unsigned otherRow) {
+void MemoryAccessMatrix::swapRows(size_t row, size_t otherRow) {
   assert((row < nRows && otherRow < nRows) && "out of bounds");
   if (row == otherRow)
     return;
-  for (unsigned col = 0; col < nColumns; ++col)
+  for (size_t col = 0; col < nColumns; ++col)
     std::swap(at(row, col), at(otherRow, col));
 }
 
-void MemoryAccessMatrix::swapColumns(unsigned column, unsigned otherColumn) {
+void MemoryAccessMatrix::swapColumns(size_t column, size_t otherColumn) {
   assert((column < nColumns && otherColumn < nColumns) && "out of bounds");
   if (column == otherColumn)
     return;
-  for (unsigned row = 0; row < nRows; ++row)
+  for (size_t row = 0; row < nRows; ++row)
     std::swap(at(row, column), at(row, otherColumn));
 }
 
-SmallVector<Value> MemoryAccessMatrix::getRow(unsigned row) const {
+SmallVector<Value> MemoryAccessMatrix::getRow(size_t row) const {
   assert(row < nRows && "the matrix must contain the given row");
   SmallVector<Value> rowCopy;
   rowCopy.reserve(nColumns);
-  for (unsigned col = 0; col < nColumns; ++col)
+  for (size_t col = 0; col < nColumns; ++col)
     rowCopy.emplace_back(at(row, col));
   return rowCopy;
 }
 
-void MemoryAccessMatrix::setRow(unsigned row, ArrayRef<Value> elems) {
+void MemoryAccessMatrix::setRow(size_t row, ArrayRef<Value> elems) {
   assert(row < nRows && "the matrix must contain the given row");
   assert(elems.size() == nColumns && "elems size must match row length!");
-  for (unsigned col = 0; col < nColumns; ++col)
+  for (size_t col = 0; col < nColumns; ++col)
     at(row, col) = elems[col];
 }
 
-void MemoryAccessMatrix::fillRow(unsigned row, Value val) {
+void MemoryAccessMatrix::fillRow(size_t row, Value val) {
   assert(row < nRows && "the matrix must contain the given row");
-  for (unsigned col = 0; col < nColumns; ++col)
+  for (size_t col = 0; col < nColumns; ++col)
     at(row, col) = val;
 }
 
-unsigned MemoryAccessMatrix::appendRow() {
+size_t MemoryAccessMatrix::appendRow() {
   ++nRows;
   data.resize(nRows * nColumns);
   return nRows - 1;
 }
 
-unsigned MemoryAccessMatrix::appendRow(ArrayRef<Value> elems) {
-  unsigned row = appendRow();
+size_t MemoryAccessMatrix::appendRow(ArrayRef<Value> elems) {
+  size_t row = appendRow();
   setRow(row, elems);
   return row;
 }
 
-SmallVector<Value> MemoryAccessMatrix::getColumn(unsigned column) const {
+SmallVector<Value> MemoryAccessMatrix::getColumn(size_t column) const {
   assert(column < nColumns && "the matrix must contain the given column");
   SmallVector<Value> columnCopy;
   columnCopy.reserve(nRows);
-  for (unsigned row = 0; row < nRows; ++row)
+  for (size_t row = 0; row < nRows; ++row)
     columnCopy.emplace_back(at(row, column));
   return columnCopy;
 }
 
-void MemoryAccessMatrix::setColumn(unsigned col, ArrayRef<Value> elems) {
+void MemoryAccessMatrix::setColumn(size_t col, ArrayRef<Value> elems) {
   assert(col < nColumns && "the matrix must contain the given column");
   assert(elems.size() == nRows && "elems size must match column length!");
-  for (unsigned row = 0; row < nRows; ++row)
+  for (size_t row = 0; row < nRows; ++row)
     at(row, col) = elems[row];
 }
 
-void MemoryAccessMatrix::fillColumn(unsigned col, Value val) {
+void MemoryAccessMatrix::fillColumn(size_t col, Value val) {
   assert(col < nColumns && "the matrix must contain the given column");
-  for (unsigned row = 0; row < nRows; ++row)
+  for (size_t row = 0; row < nRows; ++row)
     at(row, col) = val;
 }
 
 void MemoryAccessMatrix::fill(Value val) {
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col)
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col)
       at(row, col) = val;
 }
 
-MemoryAccessMatrix MemoryAccessMatrix::getRows(std::set<unsigned> rows) const {
-  assert(llvm::all_of(rows, [this](unsigned row) { return row < nRows; }) &&
+MemoryAccessMatrix MemoryAccessMatrix::getRows(std::set<size_t> rows) const {
+  assert(llvm::all_of(rows, [this](size_t row) { return row < nRows; }) &&
          "out of bounds");
   MemoryAccessMatrix subMatrix(rows.size(), nColumns);
-  unsigned pos = 0;
-  for (unsigned rowNum : rows)
+  size_t pos = 0;
+  for (size_t rowNum : rows)
     subMatrix.setRow(pos++, getRow(rowNum));
   return subMatrix;
 }
 
 MemoryAccessMatrix
-MemoryAccessMatrix::getColumns(std::set<unsigned> columns) const {
+MemoryAccessMatrix::getColumns(std::set<size_t> columns) const {
   assert(llvm::all_of(columns,
-                      [this](unsigned column) { return column < nColumns; }) &&
+                      [this](size_t column) { return column < nColumns; }) &&
          "out of bounds");
   MemoryAccessMatrix subMatrix(nRows, columns.size());
-  unsigned pos = 0;
-  for (unsigned colNum : columns)
+  size_t pos = 0;
+  for (size_t colNum : columns)
     subMatrix.setColumn(pos++, getColumn(colNum));
   return subMatrix;
 }
 
 MemoryAccessMatrix
-MemoryAccessMatrix::getSubMatrix(std::set<unsigned> rows,
-                                 std::set<unsigned> columns) const {
+MemoryAccessMatrix::getSubMatrix(std::set<size_t> rows,
+                                 std::set<size_t> columns) const {
   return getRows(std::move(rows)).getColumns(std::move(columns));
 }
 
@@ -220,8 +220,8 @@ bool MemoryAccessMatrix::isDiagonal(DataFlowSolver &solver) const {
   if (!isSquare())
     return false;
 
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isOnDiagonal = (row == col);
 
@@ -240,8 +240,8 @@ bool MemoryAccessMatrix::isIdentity(DataFlowSolver &solver) const {
   if (!isSquare())
     return false;
 
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isOnDiagonal = (row == col);
 
@@ -262,8 +262,8 @@ bool MemoryAccessMatrix::isLowerTriangular(DataFlowSolver &solver) const {
   if (!isSquare())
     return false;
 
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isAboveDiagonal = (col > row);
 
@@ -282,8 +282,8 @@ bool MemoryAccessMatrix::isUpperTriangular(DataFlowSolver &solver) const {
   if (!isSquare())
     return false;
 
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isBelowDiagonal = (col < row);
 
@@ -309,8 +309,8 @@ bool MemoryAccessMatrix::hasReverseLinearAccessPattern(
 
   // Ensure the matrix is diagonal with all non-zero elements equal to one
   // except the last one which must be equal to negative one.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isOnDiagonal = (col == row);
 
@@ -336,8 +336,8 @@ bool MemoryAccessMatrix::hasLinearOverlappedAccessPattern(
 
   // Ensure the matrix is lower triangular with all non-zero elements equal to
   // one.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isAboveDiagonal = (col > row);
 
@@ -358,8 +358,8 @@ bool MemoryAccessMatrix::hasReverseLinearOverlappedAccessPattern(
   // Ensure the matrix is lower triangular with all non-zero elements equal to
   // one except the last one on the diagonal which must be equal to negative
   // one.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isAboveDiagonal = (col > row);
 
@@ -384,8 +384,8 @@ bool MemoryAccessMatrix::hasStridedAccessPattern(DataFlowSolver &solver) const {
 
   // Ensure the matrix is diagonal with all elements equal to one except the
   // last one which must be greater than one.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isOnDiagonal = (col == row);
 
@@ -411,8 +411,8 @@ bool MemoryAccessMatrix::hasReverseStridedAccessPattern(
 
   // Ensure the matrix is diagonal with all elements equal to one except the
   // last one which must be smaller than negative one.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isOnDiagonal = (col == row);
 
@@ -438,8 +438,8 @@ bool MemoryAccessMatrix::hasStridedOverlappedAccessPattern(
 
   // Ensure the matrix is lower triangular with all non-zero elements equal to
   // one except the last one on the diagonal which must be strictly positive.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isAboveDiagonal = (col > row);
 
@@ -466,8 +466,8 @@ bool MemoryAccessMatrix::hasReverseStridedOverlappedAccessPattern(
   // Ensure the matrix is lower triangular with all non-zero elements equal to
   // one except the last one on the diagonal which must be smaller than
   // negative one.
-  for (unsigned row = 0; row < nRows; ++row)
-    for (unsigned col = 0; col < nColumns; ++col) {
+  for (size_t row = 0; row < nRows; ++row)
+    for (size_t col = 0; col < nColumns; ++col) {
       Value val = at(row, col);
       bool isAboveDiagonal = (col > row);
 
@@ -487,7 +487,7 @@ bool MemoryAccessMatrix::hasReverseStridedOverlappedAccessPattern(
 }
 
 Optional<APInt>
-MemoryAccessMatrix::getConstIntegerValue(unsigned row, unsigned column,
+MemoryAccessMatrix::getConstIntegerValue(size_t row, size_t column,
                                          DataFlowSolver &solver) const {
   Value val = at(row, column);
   return ::getConstIntegerValue(val, solver);
@@ -497,7 +497,7 @@ MemoryAccessMatrix::getConstIntegerValue(unsigned row, unsigned column,
 // OffsetVector
 //===----------------------------------------------------------------------===//
 
-OffsetVector::OffsetVector(unsigned nRows) : nRows(nRows), offsets(nRows) {}
+OffsetVector::OffsetVector(size_t nRows) : nRows(nRows), offsets(nRows) {}
 
 OffsetVector::OffsetVector(std::initializer_list<Value> initList) {
   assert(initList.size() != 0 && "Expecting a non-empty initializer list");
@@ -505,26 +505,26 @@ OffsetVector::OffsetVector(std::initializer_list<Value> initList) {
   offsets.append(initList);
 }
 
-void OffsetVector::swapRows(unsigned row, unsigned otherRow) {
+void OffsetVector::swapRows(size_t row, size_t otherRow) {
   assert((row < nRows && otherRow < nRows) && "out of bounds");
   if (row == otherRow)
     return;
   std::swap(at(row), at(otherRow));
 }
 
-Value OffsetVector::getOffset(unsigned row) const { return at(row); }
+Value OffsetVector::getOffset(size_t row) const { return at(row); }
 
-void OffsetVector::setOffset(unsigned row, Value offset) { at(row) = offset; }
+void OffsetVector::setOffset(size_t row, Value offset) { at(row) = offset; }
 
 void OffsetVector::fill(Value val) {
-  for (unsigned row = 0; row < nRows; ++row)
+  for (size_t row = 0; row < nRows; ++row)
     at(row) = val;
 }
 
-unsigned OffsetVector::append(Value offset) {
+size_t OffsetVector::append(Value offset) {
   ++nRows;
   offsets.resize(nRows);
-  unsigned lastRow = nRows - 1;
+  size_t lastRow = nRows - 1;
   setOffset(lastRow, offset);
   return lastRow;
 }
@@ -536,8 +536,8 @@ bool OffsetVector::isZero(DataFlowSolver &solver) const {
 
 bool OffsetVector::isZeroWithLastElementStrictlyPositive(
     DataFlowSolver &solver) const {
-  unsigned lastIndex = nRows - 1;
-  for (unsigned pos = 0; pos < nRows; ++pos) {
+  size_t lastIndex = nRows - 1;
+  for (size_t pos = 0; pos < nRows; ++pos) {
     Value val = at(pos);
     bool isLastIndex = (pos == lastIndex);
     if (!isLastIndex && !::isZero(val, solver))
@@ -552,8 +552,8 @@ bool OffsetVector::isZeroWithLastElementEqualTo(int k,
                                                 DataFlowSolver &solver) const {
   LLVM_DEBUG(llvm::dbgs() << "In isZeroWithLastElementEqualTo\n");
 
-  unsigned lastIndex = nRows - 1;
-  for (unsigned pos = 0; pos < nRows; ++pos) {
+  size_t lastIndex = nRows - 1;
+  for (size_t pos = 0; pos < nRows; ++pos) {
     Value val = at(pos);
     bool isLastIndex = (pos == lastIndex);
     if (!isLastIndex && !::isZero(val, solver))
@@ -566,7 +566,7 @@ bool OffsetVector::isZeroWithLastElementEqualTo(int k,
 }
 
 Optional<APInt>
-OffsetVector::getConstIntegerValue(unsigned row, DataFlowSolver &solver) const {
+OffsetVector::getConstIntegerValue(size_t row, DataFlowSolver &solver) const {
   Value val = at(row);
   return ::getConstIntegerValue(val, solver);
 }
